feat: BBTextToolTipView, BBThumbnailToolTipView 모듈 분리 - BBBaseToolTIpView 내부 drawable Method 추가 - BBToolTip Class 추가 - BBDrawable protocol, extension 제거
