+++ 14th-team5-iOS/Core/Sources/Bibbi/BBCommons/BBToolTip/BBAnimatable.swift
@@ -22,36 +22,48 @@ public protocol BBComponentShowable {
 /// **Animate**, **CGAffineTransform**, **CABasicAnimation** 을 활용한 Animation 메서드를 정의하는 Protocol 입니다.
 ///  해당 **BBComponentClosable** 프로토콜은 Component 객체를 숨기는 애니메이션을 정의하는 프로토콜입니다.
 public protocol BBComponentClosable {
-    func hidePopover(duration: TimeInterval, options: UIView.AnimationOptions, transform: CGAffineTransform, alpha: CGFloat)
+    func hide(duration: TimeInterval, options: UIView.AnimationOptions, transform: CGAffineTransform, alpha: CGFloat)
 }
 
 
 //MARK: - Extensions
 
 public extension BBComponentShowable where Self: BBToolTip {
-    /// showPopover  메서드 호출 시 Popover 애니메이션 효과를 실행합니다.
     func showPopover(duration: TimeInterval = 0.3, options: UIView.AnimationOptions = [.curveEaseInOut], transform: CGAffineTransform = CGAffineTransform(scaleX: 0.1, y: 0.1), alpha: CGFloat = 1) {
-        self.contentView.transform = transform
-        self.contentView.alpha = alpha
+        
+        guard let contentView else {
+            fatalError("contentView is Not Created")
+        }
+        
+        superview?.addSubview(contentView)
+        updateLayout()
         
         UIView.animate(withDuration: duration, delay: 0, options: options) { [weak self] in
             guard let self else { return }
-            self.contentView.transform = CGAffineTransform.identity
-            self.contentView.alpha = 1
+            self.contentView?.transform = CGAffineTransform.identity
+            self.contentView?.alpha = 1
         }
     }
 }
 
 public extension BBComponentClosable where Self: BBToolTip {
-    /// hidePopover  메서드 호출 시 Popover 애니메이션 효과를 제거합니다.
-    func hidePopover(duration: TimeInterval = 0.3, options: UIView.AnimationOptions = [.curveEaseInOut], transform: CGAffineTransform = CGAffineTransform(scaleX: 0.1, y: 0.1), alpha: CGFloat = 0) {
+    func hide(
+        duration: TimeInterval = 0.3,
+        options: UIView.AnimationOptions = [.curveEaseInOut],
+        transform: CGAffineTransform = CGAffineTransform(scaleX: 0.1, y: 0.1),
+        alpha: CGFloat = 0
+    ) {
         
-        UIView.animate(withDuration: duration, delay: 0, options: options) { [weak self] in
-            guard let self else { return }
-            self.contentView.transform = transform
-            self.contentView.alpha = alpha
-        } completion: { _ in
-            self.contentView.transform = CGAffineTransform.identity
-        }
+        UIView.animate(withDuration: duration, delay: 0, options: options, animations: { [weak self] in
+            guard let self = self else { return }
+            self.contentView?.transform = transform
+            self.contentView?.alpha = 0
+        }, completion: { [weak self] _ in
+            self?.contentView?.removeFromSuperview()
+            self?.contentView?.transform = .identity
+        })
+        
+        print("hide popover: \(self.contentView)")
+        self.contentView?.layoutIfNeeded()
     }
 }
